Refactor check_commit_opcode to RegisterTypes
