+++ synthesizer/src/process/stack/helpers/initialize.rs
@@ -114,96 +114,3 @@ impl<N: Network> Stack<N> {
         Ok(())
     }
 }
-
-impl<N: Network> RegisterTypes<N> {
-    /// Ensures the opcode is a valid opcode and corresponds to the `commit` instruction.
-    #[inline]
-    pub(crate) fn check_commit_opcode(opcode: &str, instruction: &Instruction<N>) -> Result<()> {
-        check_commit_opcode(opcode, instruction)
-    }
-}
-
-impl<N: Network> FinalizeTypes<N> {
-    /// Ensures the opcode is a valid opcode and corresponds to the `commit` instruction.
-    #[inline]
-    pub(crate) fn check_commit_opcode(opcode: &str, instruction: &Instruction<N>) -> Result<()> {
-        check_commit_opcode(opcode, instruction)
-    }
-}
-
-/// Ensures the opcode is a valid opcode and corresponds to the `commit` instruction.
-#[inline]
-fn check_commit_opcode<N: Network>(opcode: &str, instruction: &Instruction<N>) -> Result<()> {
-    // Ensure the instruction belongs to the defined set.
-    if ![
-        "commit.bhp256",
-        "commit.bhp512",
-        "commit.bhp768",
-        "commit.bhp1024",
-        "commit.ped64",
-        "commit.ped128",
-        "commit_to_group.bhp256",
-        "commit_to_group.bhp512",
-        "commit_to_group.bhp768",
-        "commit_to_group.bhp1024",
-        "commit_to_group.ped64",
-        "commit_to_group.ped128",
-    ]
-    .contains(&opcode)
-    {
-        bail!("Instruction '{instruction}' is not for opcode '{opcode}'.");
-    }
-    // Ensure the instruction is the correct one.
-    match opcode {
-        "commit.bhp256" => ensure!(
-            matches!(instruction, Instruction::CommitBHP256(..)),
-            "Instruction '{instruction}' is not for opcode '{opcode}'."
-        ),
-        "commit.bhp512" => ensure!(
-            matches!(instruction, Instruction::CommitBHP512(..)),
-            "Instruction '{instruction}' is not for opcode '{opcode}'."
-        ),
-        "commit.bhp768" => ensure!(
-            matches!(instruction, Instruction::CommitBHP768(..)),
-            "Instruction '{instruction}' is not for opcode '{opcode}'."
-        ),
-        "commit.bhp1024" => ensure!(
-            matches!(instruction, Instruction::CommitBHP1024(..)),
-            "Instruction '{instruction}' is not for opcode '{opcode}'."
-        ),
-        "commit.ped64" => ensure!(
-            matches!(instruction, Instruction::CommitPED64(..)),
-            "Instruction '{instruction}' is not for opcode '{opcode}'."
-        ),
-        "commit.ped128" => ensure!(
-            matches!(instruction, Instruction::CommitPED128(..)),
-            "Instruction '{instruction}' is not for opcode '{opcode}'."
-        ),
-        "commit_to_group.bhp256" => ensure!(
-            matches!(instruction, Instruction::CommitToGroupBHP256(..)),
-            "Instruction '{instruction}' is not for opcode '{opcode}'."
-        ),
-        "commit_to_group.bhp512" => ensure!(
-            matches!(instruction, Instruction::CommitToGroupBHP512(..)),
-            "Instruction '{instruction}' is not for opcode '{opcode}'."
-        ),
-        "commit_to_group.bhp768" => ensure!(
-            matches!(instruction, Instruction::CommitToGroupBHP768(..)),
-            "Instruction '{instruction}' is not for opcode '{opcode}'."
-        ),
-        "commit_to_group.bhp1024" => ensure!(
-            matches!(instruction, Instruction::CommitToGroupBHP1024(..)),
-            "Instruction '{instruction}' is not for opcode '{opcode}'."
-        ),
-        "commit_to_group.ped64" => ensure!(
-            matches!(instruction, Instruction::CommitToGroupPED64(..)),
-            "Instruction '{instruction}' is not for opcode '{opcode}'."
-        ),
-        "commit_to_group.ped128" => ensure!(
-            matches!(instruction, Instruction::CommitToGroupPED128(..)),
-            "Instruction '{instruction}' is not for opcode '{opcode}'."
-        ),
-        _ => bail!("Instruction '{instruction}' is not for opcode '{opcode}'."),
-    }
-    Ok(())
-}
